Bump minimal version to webpack 5

+++ test/integration/multi-hmr.test.js
@@ -1,14 +1,13 @@
 const { join } = require('node:path');
 const fs = require('node:fs');
-const timers = require('node:timers/promises');
-const { logReader, waitFor } = require('../helpers/logs.js');
 
 const execa = require('execa');
-const { test, expect, beforeEach, rstest } = require('@rstest/core');
+const { test, expect, beforeEach, afterEach, rstest } = require('@rstest/core');
+
+const { logReader, waitFor } = require('../helpers/logs.js');
+const { startBrowser, stopBrowser } = require('../helpers/puppeteer');
 
 rstest.setConfig({ testTimeout: 25_000 });
-
-const { startBrowser, stopBrowser } = require('../helpers/puppeteer');
 
 let page, util;
 beforeEach(async () => {
@@ -21,41 +20,51 @@ beforeEach(async () => {
   };
 });
 
+let fixturePath, proc;
+afterEach(async () => {
+  if (proc) {
+    proc.kill('SIGTERM');
+  }
+
+  if (fixturePath) {
+    await fs.promises.rm(fixturePath, { recursive: true, force: true });
+  }
+});
+
 test('multi compiler', async () => {
   const { getPort, replace, setup } = util;
-  const fixturePath = await setup('multi', 'multi-hmr');
-  const proc = execa('wp', [], { cwd: fixturePath });
+  fixturePath = await setup('multi', 'multi-hmr');
+  proc = execa('wp', [], { cwd: fixturePath });
   const errReader = logReader(proc.stderr);
-  const port = await getPort(proc.stdout);
+  const port = await getPort(logReader(proc.stdout));
   const url = `http://localhost:${port}`;
 
-  await waitFor('webpack: Hash:', errReader);
+  // As two compilers are running we need to wait for two events to finish
+  await waitFor('compiled successfully', errReader);
+  await waitFor('compiled successfully', errReader);
   await page.goto(url, {
     waitUntil: 'networkidle0',
   });
 
   const componentPath = join(fixturePath, 'component.js');
-  const workerPath = join(fixturePath, 'work.js');
   const componentContent = `const main = document.querySelector('main'); main.innerHTML = 'test';`;
-  const workerContent = `const worker = document.querySelector('#worker'); worker.innerHTML = 'test';`;
-
   await replace(componentPath, componentContent);
-  await waitFor('webpack: Hash:', errReader);
+  await waitFor('compiled successfully', errReader);
+
+  const workerPath = join(fixturePath, 'work.js');
+  const workerContent = `const worker = document.querySelector('#worker'); worker.innerHTML = 'test';`;
   await replace(workerPath, workerContent);
-  await waitFor('webpack: Hash:', errReader);
-  await timers.setTimeout(2_000);
+  await waitFor('compiled successfully', errReader);
 
-  const componentValue = await page.evaluate(
-    () => document.querySelector('main').innerHTML,
-  );
-  const workValue = await page.evaluate(
-    () => document.querySelector('#worker').innerHTML,
-  );
+  await expect
+    .poll(async () =>
+      page.evaluate(() => document.querySelector('main').innerHTML),
+    )
+    .toBe('test');
 
-  proc.kill('SIGTERM');
-
-  expect(componentValue).toEqual('test');
-  expect(workValue).toEqual('test');
-
-  await fs.promises.rm(fixturePath, { recursive: true, force: true });
+  await expect
+    .poll(async () =>
+      page.evaluate(() => document.querySelector('#worker').innerHTML),
+    )
+    .toBe('test');
 });
